+++ v3/integrations/nrfasthttp/go.mod
@@ -3,8 +3,6 @@ module github.com/newrelic/go-agent/v3/integrations/nrfasthttp
 go 1.19
 
 require (
-	github.com/newrelic/go-agent/v3 v3.26.0
-	github.com/stretchr/testify v1.8.4
-	github.com/valyala/fasthttp v1.48.0
+	github.com/newrelic/go-agent/v3 v3.28.0
+	github.com/valyala/fasthttp v1.49.0
 )
-replace github.com/newrelic/go-agent/v3 => ../..
